+++ src/views/Result/ParticipantResult.jsx
@@ -200,7 +200,7 @@ const ParticipantResult = () => {
                                             const submissionTime = new Date(result.submissionTime);
                                             const userTimeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
 
-                                            const options = { month: 'long', day: 'numeric', year: 'numeric', hour: 'numeric', minute: 'numeric', hour12: true, timeZone: userTimeZone };
+                                            const options = { month: 'short', day: 'numeric', year: 'numeric', hour: 'numeric', minute: 'numeric', hour12: true, timeZone: userTimeZone };
                                             const formattedDate = new Intl.DateTimeFormat('en-US', options).format(submissionTime);
 
                                             const truncatedText = `${formattedDate}`;
